Update custom_attributes.cpp
This is to fix a crash that can occur when a robot shoots a friendly tank with something dealing negative damage. As far as I can tell, I only got this to crash with a Righteous Bison energy ball. This crash was discovered in the Haphazard Routine mission from Operation Madness Vs. Machines, when a "Healing Ray Soldier"'s energy blast came in contact with a tank. 

https://wiki.teamfortress.com/wiki/Haphazard_Routine_(mission)#Healing_Ray_Soldier
Pictured here: https://cdn.discordapp.com/attachments/78276459206025216/1062998274165977128/image.png

I recompiled with the added code and it seemed to fix it.